Title: 20대의 평균 나이, Time: 127ms, Memory: 10MB, Status: Passed - CodeTree

+++ 250106/20대의 평균 나이/average-age-of-20.java
@@ -4,7 +4,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int sum = 0;
-        double cnt = 0;
+        int cnt = 0;
         while (true) {
             int a = sc.nextInt();
 
@@ -18,6 +18,6 @@ public class Main {
                 break;
             }   
         }
-        System.out.printf("%.2f",sum/cnt);
+        System.out.printf("%.2f",(double)sum/cnt);
     }
 }
